add validation for socio economic
