--- src/containers/Lifemap/Components/SocioEconomic/helpers/validation.js
@@ -1,7 +1,0 @@
-export const validate = values => {
-  const errors = {}
-  if (!values.reason) {
-    errors.reason = 'Required'
-  }
-  return errors
-}
